requirements.txt
Authlib==1.3.2
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Flask==2.3.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-OAuthlib==0.9.6
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-oauth2-tool==0.0.3
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
oauth2client==4.1.3
oauthlib==2.1.0
pino==0.6.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.1.0
rsa==4.9
six==1.16.0
smmap==5.0.1
soupsieve==2.6
style==1.1.6
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==2.3.8
